Feature: Added the option of switching the "explorer pane" between viewing the "toolbox" (the project's circuits and libraries that has appeared in the explorer pane in previous versions) and a simulation hierarchy.
Feature: Splitters can be configured to show which bits correspond to each end using the new Appearance attribute. It uses this new appearance by default.

Bug fix: In moving a component whose output is connected to itself, it is possible for an exception to cause the program to hang.

Design change: Removed option to disable the project toolbar.